requirements.txt
matplotlib
numpy
networkx
cirq-core
stim
stimcirq
Flask==3.0.0
flask_cors==3.0.10